igmp.c: Attempt to clean up #ifdef hell in accept_igmp()
This patch is inherited from pimd commit 54ede55.

The mess is due to various UNIX flavors handling SOCK_RAW and IP_HDRINCL
a bit differently.  It seems though that NetBSD is the only one (of the
major players) that still is a bit different from the others.  NetBSD
adjusts ip_len with the IP header size before it forwards the frame to
the userspace RAW socket.  Not so raw after all.

Since we cannot rely, in any sane way, on ip_len being correct, this
patch replaces the IGMP packet length sanity check that until now has
used `ip->ip_len` in igmp.c, with `recvlen` that pim.c already use in
`accept_pim()`.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>