pimctl: fix ipc socket read, exit too early
When the receiving client buffer is too small for data remaining at the
end of the transfer from the daemon, we get both POLLIN and POLLHUP on
the socket.  In that case we must continue reading and not process the
POLLHUP, otherwise the client will not get the last remaining data.

Also, handle EINTR in case someone presses Ctrl-C in the middle of the
transmission.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>