test: refactor, main netns shared between gcc and clang workflows
Running two workflows in parallel means they may run the exact same test
a the same time on the same machine, sharing the same main netns.  This
patch is a refactor that aims to create all veth pairs inside sub-netns
and move the other ends of each to empty netns shells before doing the
rest of the setup and test.

Note: still has annoying artifacts in output.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>